Fix minor downcasting warnings

+++ IlmBase/Imath/ImathMatrix.h
@@ -3252,13 +3252,13 @@ operator << (std::ostream &s, const Matrix33<T> &m)
     if (s.flags() & std::ios_base::fixed)
     {
         s.setf (std::ios_base::showpoint);
-        width = s.precision() + 5;
+        width = (int) s.precision() + 5;
     }
     else
     {
         s.setf (std::ios_base::scientific);
         s.setf (std::ios_base::showpoint);
-        width = s.precision() + 8;
+        width = (int) s.precision() + 8;
     }
 
     s << "(" << std::setw (width) << m[0][0] <<
@@ -3287,13 +3287,13 @@ operator << (std::ostream &s, const Matrix44<T> &m)
     if (s.flags() & std::ios_base::fixed)
     {
         s.setf (std::ios_base::showpoint);
-        width = s.precision() + 5;
+        width = (int) s.precision() + 5;
     }
     else
     {
         s.setf (std::ios_base::scientific);
         s.setf (std::ios_base::showpoint);
-        width = s.precision() + 8;
+        width = (int) s.precision() + 8;
     }
 
     s << "(" << std::setw (width) << m[0][0] <<
