+++ IlmBase/ImathTest/testExtractSHRT.cpp
@@ -56,8 +56,8 @@ using namespace IMATH_INTERNAL_NAMESPACE;
 
 namespace {
 
-float rad (float deg) {return deg * (M_PI / 180);}
-float deg (float rad) {return rad * (180 / M_PI);}
+float rad (float deg) {return deg * (float) (M_PI / 180);}
+float deg (float rad) {return rad * (float) (180 / M_PI);}
 
 
 void
@@ -130,8 +130,8 @@ testRandomAngles33 ()
 	// Scale M.
 	//
 
-	V2f s (random.nextf (0.000001, 2.0),
-	       random.nextf (0.000001, 2.0));
+	V2f s ((float) random.nextf (0.000001, 2.0),
+	       (float) random.nextf (0.000001, 2.0));
 
 	for (int j=0; j < 2; j++)
 	    if (random.nextf (0.0, 1.0) >= 0.5)
@@ -143,7 +143,7 @@ testRandomAngles33 ()
 	// Shear M.
 	//
 
-	float h = random.nextf (0.000001, 2.);
+	float h = (float) random.nextf (0.000001, 2.);
 	if (random.nextf (0.0, 1.0) >= 0.5)
 	    h *= -1;
 
@@ -153,7 +153,7 @@ testRandomAngles33 ()
 	// Rotate M.
 	//
 
-	float r = rad (random.nextf (-180, 180));
+	float r = rad ((float) random.nextf (-180, 180));
 	
 	M *= M33f().setRotation (r);
 
@@ -161,8 +161,8 @@ testRandomAngles33 ()
 	// Translate M.
 	//
 
-	V2f t (random.nextf (-10, 10), 
-	       random.nextf (-10, 10));
+	V2f t ((float) random.nextf (-10, 10), 
+	       (float) random.nextf (-10, 10));
 	
 	M *= M33f().setTranslation (t);
 
@@ -293,9 +293,9 @@ testRandomAngles44 ()
 	// Translate M.
 	//
 
-	V3f t (random.nextf (-10, 10), 
-	       random.nextf (-10, 10), 
-	       random.nextf (-10, 10));
+	V3f t ((float) random.nextf (-10, 10), 
+	       (float) random.nextf (-10, 10), 
+	       (float) random.nextf (-10, 10));
 	
 	M.translate (t);
 
@@ -303,9 +303,9 @@ testRandomAngles44 ()
 	// Rotate M.
 	//
 
-	V3f r (rad (random.nextf (-180, 180)),
-	       rad (random.nextf (-180, 180)),
-	       rad (random.nextf (-180, 180)));
+	V3f r (rad ((float) random.nextf (-180, 180)),
+	       rad ((float) random.nextf (-180, 180)),
+	       rad ((float) random.nextf (-180, 180)));
 	
 	M.rotate (r);
 
@@ -313,9 +313,9 @@ testRandomAngles44 ()
 	// Shear M.
 	//
 
-	V3f h (random.nextf (0.000001, 2.0), 
-	       random.nextf (0.000001, 2.0), 
-	       random.nextf (0.000001, 2.0));
+	V3f h ((float) random.nextf (0.000001, 2.0), 
+	       (float) random.nextf (0.000001, 2.0), 
+	       (float) random.nextf (0.000001, 2.0));
 	
 	for (int j=0; j < 3; j++)
 	    if (random.nextf (0.0, 1.0) >= 0.5)
@@ -327,9 +327,9 @@ testRandomAngles44 ()
 	// Scale M.
 	//
 
-	V3f s (random.nextf (0.000001, 2.0),
-	       random.nextf (0.000001, 2.0),
-	       random.nextf (0.000001, 2.0));
+	V3f s ((float) random.nextf (0.000001, 2.0),
+	       (float) random.nextf (0.000001, 2.0),
+	       (float) random.nextf (0.000001, 2.0));
 
 	for (int j=0; j < 3; j++)
 	    if (random.nextf (0.0, 1.0) >= 0.5)
@@ -419,7 +419,7 @@ test ()
     for (int i = 0; i < 360; i += 90)
 	for (int j = 0; j < 360; j += 90)
 	    for (int k = 0; k < 360; k += 90)
-		testAngles44 (V3f (i, j, k));
+		testAngles44 (V3f ((float) i, (float) j, (float) k));
 }
